fix: update ut, telemetry and move helper function to common
