+++ packages/fx-core/src/component/driver/aad/utility/constants.ts
@@ -56,4 +56,6 @@ export const constants = {
 
 export const telemetryKeys = {
   newAadApp: "new-aad-app",
+  userInputAadApp: "user-input-aad-app",
+  insufficientPermissionAadApp: "insufficient-permission-aad-app",
 };
